backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlmodel==0.0.16
passlib==1.7.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9 